#999: Namespaces fixes
