+++ .NET/OneBeyond.DataAccess/Author/IAuthorRepository.cs
@@ -1,0 +1,9 @@
+﻿using OneBeyond.Model.Entities;
+
+namespace OneBeyond.DataAccess;
+public interface IAuthorRepository
+{
+    public List<Author> GetAuthors();
+
+    public Guid AddAuthor(Author author);
+}
