+++ .NET/OneBeyond.DataAccess/Book/IBookRepository.cs
@@ -1,0 +1,10 @@
+﻿using OneBeyond.Model.Entities;
+
+namespace OneBeyond.DataAccess;
+
+public interface IBookRepository
+{
+    public List<Book> GetBooks();
+
+    public Guid AddBook(Book book);
+}
