--- .NET/OneBeyond.DataAccess/IAuthorRepository.cs
@@ -1,11 +1,0 @@
-﻿using OneBeyondApi.Model;
-
-namespace OneBeyondApi.DataAccess
-{
-    public interface IAuthorRepository
-    {
-        public List<Author> GetAuthors();
-
-        public Guid AddAuthor(Author author);
-    }
-}
