--- .NET/OneBeyond.DataAccess/IBookRepository.cs
@@ -1,11 +1,0 @@
-﻿using OneBeyondApi.Model;
-
-namespace OneBeyondApi.DataAccess
-{
-    public interface IBookRepository
-    {
-        public List<Book> GetBooks();
-
-        public Guid AddBook(Book book);
-    }
-}
