--- .NET/OneBeyond.DataAccess/ICatalogueRepository.cs
@@ -1,11 +1,0 @@
-﻿using OneBeyondApi.Model;
-
-namespace OneBeyondApi.DataAccess
-{
-    public interface ICatalogueRepository
-    {
-        public List<BookStock> GetCatalogue();
-
-        public List<BookStock> SearchCatalogue(CatalogueSearch search);
-    }
-}
