+++ .NET/onebeyond.Model/DTOs/BorrowerLoanDto.cs
@@ -1,3 +1,3 @@
-﻿namespace OneBeyondApi.Model.DTOs;
+﻿namespace OneBeyond.Model.DTOs;
 
 public sealed record BorrowerLoanDto(string BorrowerName, string BorrowerEmail, string BookTitle, DateTime? LoadEndDate);
