+++ .NET/onebeyond.Model/DTOs/ReservationStatusDto.cs
@@ -1,3 +1,3 @@
-﻿namespace OneBeyondApi.Model.DTOs;
+﻿namespace OneBeyond.Model.DTOs;
 
 public sealed record ReservationStatusDto(string BorrowerName, string BookName, DateTime ReservationDate, DateTime ExpectedAvailabilityDate);
