+++ .NET/onebeyond.Model/Entities/Author.cs
@@ -1,8 +1,6 @@
-﻿namespace OneBeyondApi.Model
+﻿namespace OneBeyond.Model.Entities;
+public class Author
 {
-    public class Author
-    {
-        public Guid Id { get; set; }
-        public string Name { get; set; }
-    }
+    public Guid Id { get; set; }
+    public string Name { get; set; }
 }
