+++ .NET/onebeyond.Model/Entities/BookStock.cs
@@ -1,10 +1,8 @@
-﻿namespace OneBeyondApi.Model
+﻿namespace OneBeyond.Model.Entities;
+public class BookStock
 {
-    public class BookStock
-    {
-        public Guid Id { get; set; }
-        public Book Book { get; set; }
-        public DateTime? LoanEndDate { get; set; }
-        public Borrower? OnLoanTo { get; set; }
-    }
+    public Guid Id { get; set; }
+    public Book Book { get; set; }
+    public DateTime? LoanEndDate { get; set; }
+    public Borrower? OnLoanTo { get; set; }
 }
