+++ .NET/onebeyond.Model/Entities/Borrower.cs
@@ -1,9 +1,8 @@
-﻿namespace OneBeyondApi.Model
+﻿namespace OneBeyond.Model.Entities;
+
+public class Borrower
 {
-    public class Borrower
-    {
-        public Guid Id { get; set; }
-        public string Name { get; set; }
-        public string EmailAddress { get; set; }
-    }
+    public Guid Id { get; set; }
+    public string Name { get; set; }
+    public string EmailAddress { get; set; }
 }
